+++ lib/lacerda/publish_contract.rb
@@ -13,6 +13,19 @@ module Lacerda
       @comparator.contains?(consumer.consume.scoped_schema(service), consumer.name)
     end
 
+    def object(name)
+      scoped_name = name.to_s.underscore
+
+      # Add our own prefix automatically if necessary
+      unless scoped_name.start_with?(service.name.underscore)
+        scoped_name = [service.name.underscore, scoped_name].join(Lacerda::SCOPE_SEPARATOR)
+      end
+
+      schema = @schema[:definitions][scoped_name]
+      raise Lacerda::Service::InvalidObjectTypeError.new(scoped_name) unless schema
+      Lacerda::PublishedObject.new(service, scoped_name, schema)
+    end
+
     private
 
     def object_description_class