+++ spec/lib/lacerda/service_spec.rb
@@ -59,33 +59,59 @@ describe Lacerda::Service do
   context "validating objects" do
 
     context "to publish" do
-
       it "complains about an unknokwn type" do
         expect(
-          publisher.validate_object_to_publish(:unknown_type, {some: :data})
+          publisher.validate_object_to_publish('unknown_type', {some: :data})
         ).to be false
       end
 
       it "complains about an unknokwn type with an exception" do
         expect{
-          publisher.validate_object_to_publish!(:unknown_type, {some: :data})
+          publisher.validate_object_to_publish!('unknown_type', {some: :data})
         }.to raise_error(Lacerda::Service::InvalidObjectTypeError)
       end
 
       it "accepts a valid object" do
         valid_post = {id: 1, title: 'My title'}
         expect(
-          publisher.validate_object_to_publish(:post, valid_post)
+          publisher.validate_object_to_publish('Post', valid_post)
         ).to be true
       end
 
       it "rejects an valid object with an exception" do
         invalid_post = {id: 'string', title: 'My title'}
         expect{
-          publisher.validate_object_to_publish!(:post, invalid_post)
+          publisher.validate_object_to_publish!('Post', invalid_post)
         }.to raise_error(JSON::Schema::ValidationError)
       end
+    end
 
+    context "to consume" do
+      it "complains about an unknokwn type" do
+        expect(
+          consumer.validate_object_to_consume('Publisher:unknown_type', {some: :data})
+        ).to be false
+      end
+
+      it "complains about an unknokwn type with an exception" do
+        expect{
+          consumer.validate_object_to_consume!('Publisher:unknown_type', {some: :data})
+        }.to raise_error(Lacerda::Service::InvalidObjectTypeError)
+      end
+
+      it "accepts a valid object" do
+        valid_post = {id: 1, title: 'My title'}
+        expect(
+          consumer.validate_object_to_consume('Publisher:Post', valid_post)
+        ).to be true
+      end
+
+      it "rejects an valid object with an exception" do
+        invalid_post = {tag: 'string', title: 'My title'}
+        expect{
+          consumer.validate_object_to_consume!('Publisher:Post', invalid_post)
+        }.to raise_error(JSON::Schema::ValidationError)
+      end
     end
 
   end